✨ feat(version): 更新应用版本至 2026.05.12.01
- 更新 build.gradle 中的 versionCode 和 versionName
- 更新 release/output-metadata.json 中的版本信息

♻️ refactor(offset): 重构位置偏移配置为 0.1dp 精度

- 【配置存储】将位置偏移量存储单位从 dp 升级为 0.1dp，提升调节精度
- 【配置迁移】添加自动迁移逻辑，兼容旧版 dp 配置
- 【界面交互】更新 MainActivity 中的滑块和数值输入，支持 0.1dp 精度输入
- 【备份兼容】更新配置导入导出逻辑，支持 v2 和 v3 版本配置
- 【核心计算】更新 SettingsStore 中的单位转换和精度处理函数
- 【绘图模块】更新 FlymeStatusBarSizer、IconMetrics、IosBatteryPainter、OneUiBatteryPainter、RightIconGroupPreviewView、SignalIconDrawable、SignalPreviewPainter、WifiIconDrawable 中的偏移量计算和绘图逻辑，统一使用 float 类型处理像素偏移

+++ app/src/main/java/com/example/flymestatusbarsizer/IosBatteryPainter.java
@@ -61,7 +61,7 @@ final class IosBatteryPainter {
     static void draw(Canvas canvas, Rect bounds, int level, boolean pluggedIn, boolean charging,
             boolean quickCharging,
             int fillColor, int textColor, boolean showLevelText, float textScale, Typeface typeface,
-            int bodyYOffsetPx, int textYOffsetPx, int boltYOffsetPx,
+            float bodyYOffsetPx, float textYOffsetPx, float boltYOffsetPx,
             boolean hollow, boolean hollowFillFollowsLevel) {
         if (bounds.width() <= 0 || bounds.height() <= 0) {
             return;
@@ -143,7 +143,7 @@ final class IosBatteryPainter {
 
     private static void drawHollowBattery(Canvas canvas, float contentRadius, float capContentRadius,
             int emptyColor, int fillColor, int level, String levelText, float textSize,
-            int textYOffsetPx, boolean showLevelText, boolean showBolt, int boltColor,
+            float textYOffsetPx, boolean showLevelText, boolean showBolt, int boltColor,
             boolean quickCharging,
             boolean fillFollowsLevel) {
         if (!showLevelText) {
@@ -180,8 +180,8 @@ final class IosBatteryPainter {
         canvas.restoreToCount(layer);
     }
 
-    private static void applyBodyVerticalOffset(int bodyYOffsetPx) {
-        if (bodyYOffsetPx == 0) {
+    private static void applyBodyVerticalOffset(float bodyYOffsetPx) {
+        if (bodyYOffsetPx == 0f) {
             return;
         }
         float offsetY = -bodyYOffsetPx;
